Store DTLS epoch state separately
This wraps DTLS per-epoch state into DTLSReadEpoch and DTLSWriteEpoch
structures. For now, we only keep one DTLSReadEpoch though we will need
to keep two in DTLS 1.3. In preparation for that, I've reworked the DTLS
record parser to resolve the DTLSReadEpoch at the same time, even though
there's currently only one of them.

On the write side, this removes the special-cased initial write epoch
and just stores an array of the last few epochs.

Some things this does not yet do, but that we ideally would do as
follow-ups:

1. Move RecordNumberEncrypter out of SSLAEADContext, now that we have a
   DTLS-specific struct for it.

2. Pass just a byte secret into set_read_state / set_write_state and let
   the transport-aware code construct the SSLAEADContext.

3. Don't construct the SSLAEADContext at all for QUIC.

4. Don't save the read and write traffic secrets in QUIC at all.

5. KeyUpdate should be part of the SSL_PROTOCOL_METHOD interface to
   accomodate DTLS driving KeyUpdate by ACK.

Update-Note: As part of rearranging the record parser, when the DTLS 1.2
implementation encounters a DTLS 1.3 record, it will now discard just
that record and continue parsing records out of the packet, rather than
discarding the whole packet. This isn't expected to make any difference.

Bug: 371998381
Change-Id: Ie2ae657d41e33152208a001df177630398798394
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/72128
Commit-Queue: David Benjamin <davidben@google.com>
Reviewed-by: Nick Harper <nharper@chromium.org>