Move `alignas` to start of declaration.
Prior to Clang 19, the `alignas` keyword was being parsed in C23 mode
(where it is newly a keyword, instead of a define) using the rules for
C++ `alignas`, rather than the slightly-different C parsing rules. It
was therefore not accepted after `static`, with `error: an attribute
list cannot appear here`.

Moving the `alignas` to the beginning of the declaration allows it to
build with all versions.

Thanks to James Knight for the change.

Change-Id: I3c109895c47fe7db185a242f310ca07d1e66f609
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/71807
Commit-Queue: Bob Beck <bbe@google.com>
Auto-Submit: Adam Langley <agl@google.com>
Commit-Queue: Adam Langley <agl@google.com>
Reviewed-by: Bob Beck <bbe@google.com>