Pull the DTLS reassembly bitmap into its own abstraction
We'll use the same bitmap for tracking which parts of a message have
been ACKed. To that end, I've gone ahead and implemented a
NextUnmarkedRange.

This does make the hm_fragment structure a little larger, but we
indirect them to the heap, so I think this is fine. In the steady state,
they do not contribute to per-connection memory use.

Bug: 42290594
Change-Id: I326520454ee4d6832248b50ea3f5205f41d74cbe
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/72269
Auto-Submit: David Benjamin <davidben@google.com>
Reviewed-by: Nick Harper <nharper@chromium.org>
Commit-Queue: David Benjamin <davidben@google.com>