util/fipstools: add EDDSA ACVP support
This commit extends the acvptool subprocess package to support the
EDDSA test vectors and expected responses defined by
draft-celi-acvp-eddsa:

https://pages.nist.gov/ACVP/draft-celi-acvp-eddsa.html

There was a pre-existing registration for the "EDDSA" command, but it
was using the ecdsa.go subprocess code that isn't a match to the
EDDSA specification. No EDDSA capabilities were being claimed in the
modulewrapper.cc or testmodulewrapper.go JSON so I suspect this was
perhaps the start of unfinished support (?).

There are a few quirks for EDDSA compared to ECDSA beyond the
differences in vectors/expected results. Notably there is the option to
advertise the prehashed capability to exercise Ed25519ph.

To facilitate the testmodulewrapper being able to implement EDDSA keyVer
there's a new dependency added on Filippo's edwards25519 package. The go
stdlib equivalent is not exported for external use.

Change-Id: Id0571470202a2761125cc9af3c5f0593b0b94ccc
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/72207
Commit-Queue: Bob Beck <bbe@google.com>
Reviewed-by: Bob Beck <bbe@google.com>
Reviewed-by: Adam Langley <agl@google.com>