Never reconstruct 1.3 record numbers above 2^48-1
In DTLS 1.2, it was syntactically impossible to make a record number
over 2^48-1 or an epoch number over 2^16-1. This was nice because it
meant record numbers could be packed into an 8-byte structure.

DTLS 1.3 attempted to lift these constraints, but ineffectively. The
epoch count cannot go past 2^16-1 due to how the message number works.
The outgoing record count has no reason to exceed 2^48-1 because we
would have to send 5 petabytes to count that high even with empty
records, and indeed we check this right now.

The incoming record count is complicated. Record number reconstruction
*could* count that high if we receive 2^33 records (the max we can jump
is 2^15 at a time). This is within the realm of possibility for an
attacker, but too high to practically count via unit tests, and useless
for a well-behaved sender because the sender has no reason to skip
sequence numbers.

The result is that we need to use more memory and worry about a case
that we cannot practically test, which is only useful for attackers.
This is not useful. Instead, change the record number reconstruction
logic to never give back record numbers past 2^48. If the peer manages
to send 5 petabytes and count that high without rekeying, we'll jam the
record number and start dropping those packets. An error would be
preferable, but that would be a codepath we cannot effectively test, so
I opted to just bodge the reconstruction logic.

With this done, we can pack the record numbers again. The next CL will
introduce a DTLSRecordNumber type to abstract the various bit
manipulations we've been doing.

Bug: 42290594
Change-Id: I8bb0584a792578cf1d2a9df0b0c035d794c85beb
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/72270
Auto-Submit: David Benjamin <davidben@google.com>
Reviewed-by: Nick Harper <nharper@chromium.org>
Commit-Queue: Nick Harper <nharper@chromium.org>