Return the record number out of the DTLS record layer
For now this is ignored by the caller, but ACK management will require
this value. To do this, I've introduced a DTLSRecordNumber type. This
isn't quite used in all the places that we could use, so I've left a
TODO about this.

As part of this, also fix some of the weirdness where some code would
mix up seq and epoch+seq across 1.2 and 1.3.

Fixing that up also revealed that actually we'd accidentally changed
SSL_get_read_sequence slightly in the epoch state refactor. This CL ends
up tweaking it slightly more in the DTLS 1.3 case (which doesn't work
anyway). I've updated the TODO comment accordingly.

Bug: 42290594
Change-Id: Ie7e7da02c432a137ae2314fb3557d1c3717952f3
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/72271
Reviewed-by: Nick Harper <nharper@chromium.org>
Commit-Queue: David Benjamin <davidben@google.com>