Store an Array in hm_fragment
Also rename it to DTLSIncomingMessage to mirror DTLSOutgoingMessage.
(Also renamed to be C++ style like other, slightly more C++-y,
internals.)

It is slightly less compact in memory to store the size as a size_t
rather than a uint32_t that can be packed in there, but this object is
largely discarded after the handshake, and this way we get more bounds
checking.

I haven't actually replaced the memcpy with a span-based copy since we
don't have one yet, but this opens the door for that.

Bug: 374890768
Change-Id: I0f69fc4ca64bebed95d90979ec99fe1ea57c66fb
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/72273
Reviewed-by: Nick Harper <nharper@chromium.org>
Commit-Queue: David Benjamin <davidben@google.com>