Amortize invariant maintenance in DTLSMessageBitmap
After we receive a handshake fragment, we check if the fragment is now
complete. This requires scanning over the entire message bitmap. This is
the one step in MarkRange that does not scale with the number of bytes
the peer sent us.

This is not really a DoS vector due to message size limits but, in
principle, the peer could send us all but the last byte of a message and
then repeatedly send us the first byte, causing us to scan over
msg_len / 8 bytes each time. We can easily amortize this by saving how
far we scanned last time, because bits will never be unmarked. This
means we only advance over each byte in the bitmap once.

Change-Id: I802003d587de4c0a45650b891c76cea8b10acd17
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/72287
Reviewed-by: Nick Harper <nharper@chromium.org>
Commit-Queue: David Benjamin <davidben@google.com>