Implement a fixed-width ring buffer for DTLS 1.3
DTLS 1.3 ACK management requires that we track:

- Received record numbers, to send in ACKs

- Information about unacked outgoing records, so we can update message
  state when ACKs from in

Both of those lists can, in principle, grow unboundedly, but can also be
freely dropped. This class is intended as a fixed-width ring buffer for
managing this. I was worried I'd need need to dip into the scary
uninitialized functions again but actually this is just an InplaceVector
and an integer, so that was quite satisfying.

Bug: 374890768
Change-Id: I4037ef5875a2df70f0dc16df1bbba478362cecce
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/72274
Reviewed-by: Nick Harper <nharper@chromium.org>
Commit-Queue: David Benjamin <davidben@google.com>