Pack DTLS fragments into records when possible
We were sending at most one DTLS fragment per record, which prevents us
from amortizing the overhead. In DTLS 1.2, this doesn't impact the
encrypted epoch (just Finished), but it does make the unencrypted epoch
slightly more efficient.

In DTLS 1.3, it doesn't impact the unencrypted epochs, but the encrypted
epochs become a lot more efficient. This is the DTLS analog to
https://boringssl-review.googlesource.com/28744

As part of this, test that:

- Each handshake fragment contains as much of the message as can fit in
  the packet.

- If the first record of packet N could have (partially) fit in packet
  N-1, the caller should have put it in packet N-1.

- We never have two handshake records of the same epoch in a row in one
  packet. If we did, those could have been merged for less overhead.
  (That's what's fixed by this CL.)

As part of this, we can remove the MTUExceeded tests. Those were, badly,
testing that the peer successfully filled at least one packet. Now we
test packet-filling more directly.

Finally, although we don't particularly care about CBC ciphers, I've
made the packet-filling code more accurate for CBC ciphers. This is
purely so that we don't need as many special cases on the runner side
when deciding if there was more room available in the packet.

Bug: 42290594
Change-Id: Ia79b2a09d3008e95de08121fc4e35768c6f3ae77
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/72275
Auto-Submit: David Benjamin <davidben@google.com>
Commit-Queue: David Benjamin <davidben@google.com>
Reviewed-by: Nick Harper <nharper@chromium.org>