Redo DTLS retransmit tests
Previously, we tested DTLS retransmit by hooking *before* processing a
flight and bracketing one flight's worth of packets and then dropping
all of them, with minimal processing. We'd then repeat this a few times
and, once we were satisfied, actually process the next flight. This has
a few drawbacks:

1. We cannot interleave timeouts with chunks of the next flight.
   Retransmit of the peer's flight, in both 1.2 and 1.3, interacts with
   the peer receiving all or parts of the next flight, and in different
   ways.

2. We cannot easily, in 1.3, ACK part of a flight and then wait for the
   shim to retransmit the other part.

The new design instead hooks *after* we construct the reply to the
peer's flight under test. We buffer up all the data to be sent and then
call into a test-supplied callback that can control the exact order of
timeouts, etc. See the comment on DTLSController for more details.

As part of this, ChangeCipherSpec is now implicitly ordered with
ReorderHandshakeFragments, so there is no more need to carry
ReorderChangeCipherSpec separately.

Bug: 42290594
Change-Id: I754946dbb5591188aa2d32b694f213269836b266
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/72708
Reviewed-by: Nick Harper <nharper@chromium.org>
Auto-Submit: David Benjamin <davidben@google.com>
Commit-Queue: David Benjamin <davidben@google.com>