Better track final vs early versions
https://boringssl-review.googlesource.com/c/boringssl/+/71534 generally
made things cleaner, but lost a subtlety: when we check if the version
is set, sometimes we care about whether it is an early version or final
version. This includes the two minor behavior changes listed there, and
a host of TODOs we kept adding to the DTLS 1.3 implementation.

We don't yet implement DTLS 1.3 0-RTT, but let's just resolve these
TODOs now rather than constantly having to make a note of this.

Bug: 42290594
Change-Id: I74e1367836762a6e6fd8d216158ee5309d387cce
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/72887
Reviewed-by: Nick Harper <nharper@chromium.org>
Commit-Queue: David Benjamin <davidben@google.com>